sinkv2(ticdc): better code
